refactor(formatter): remove `@ts-ignore` for `DateTime`
